+++ tests/testthat/test-fileDepth.R
@@ -1,0 +1,8 @@
+context("fileDepth")
+
+test_that("fileDepth", {
+    expect_identical(
+        object = fileDepth(file.path("", "tmp")),
+        expected = 1L
+    )
+})