+++ NEWS.md
@@ -3,6 +3,8 @@
 - Migrated `realpath` from brio package. Reworked slightly to not depend on any
   assert checks from goalie package. Now this function essentially wraps
   `normalizePath` but provides a more informative error message.
+- `basenameSansExt`: Minor tweak to ensure that files without an extension still
+  return basename as expected.
 
 ## acidbase 0.1.0 (2019-10-21)